put the database initialization in a different file
